Fix #299: Missing closing tag in XML output
When there is at least one invoked method, a div for the chronological-class is
opened on line 75. This div is closed as long as there are subsequent classes
(by the call to pop on line 73), however, for the last class no corresponding pop
call occurs. The fix closes this tag, if it was opened.